style: Kill StoredRestyleHint.
Bug: 1368236
MozReview-Commit-ID: 43Cf0rJyhzO
